refactored LunchService to separate repository logic into repository package
