+++ src/main/java/com/rezdy/lunch/repository/RecipeRepositoryImpl.java
@@ -1,0 +1,37 @@
+package com.rezdy.lunch.repository;
+
+import com.rezdy.lunch.service.Recipe;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+
+import javax.persistence.EntityManager;
+import javax.persistence.criteria.*;
+import java.time.LocalDate;
+import java.util.List;
+
+@Repository
+public class RecipeRepositoryImpl implements RecipeRepository {
+
+    @Autowired
+    private EntityManager entityManager;
+
+    public List<Recipe> loadRecipes(LocalDate date) {
+        CriteriaBuilder cb = entityManager.getCriteriaBuilder();
+        CriteriaQuery<Recipe> criteriaQuery = cb.createQuery(Recipe.class);
+        Root<Recipe> recipeRoot = criteriaQuery.from(Recipe.class);
+
+        CriteriaQuery<Recipe> query = criteriaQuery.select(recipeRoot);
+
+        Subquery<Recipe> nonExpiredIngredientSubquery = query.subquery(Recipe.class);
+        Root<Recipe> nonExpiredIngredient = nonExpiredIngredientSubquery.from(Recipe.class);
+        nonExpiredIngredientSubquery.select(nonExpiredIngredient);
+
+        Predicate matchingRecipe = cb.equal(nonExpiredIngredient.get("title"), recipeRoot.get("title"));
+        Predicate expiredIngredient = cb.lessThan(nonExpiredIngredient.join("ingredients").get("useBy"), date);
+
+        Predicate allNonExpiredIngredients = cb.exists(nonExpiredIngredientSubquery.where(matchingRecipe, expiredIngredient));
+
+        return entityManager.createQuery(query.where(allNonExpiredIngredients)).getResultList();
+    }
+}
+
