+++ src/main/java/com/rezdy/lunch/service/LunchService.java
@@ -1,5 +1,6 @@
 package com.rezdy.lunch.service;
 
+import com.rezdy.lunch.repository.RecipeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,12 +14,12 @@ import java.util.List;
 public class LunchService {
 
     @Autowired
-    private EntityManager entityManager;
+    private RecipeRepository recipeRepository;
 
     private List<Recipe> recipesSorted;
 
     public List<Recipe> getNonExpiredRecipesOnDate(LocalDate date) {
-        List<Recipe> recipes = loadRecipes(date);
+        List<Recipe> recipes = recipeRepository.loadRecipes(date);
 
         sortRecipes(recipes);
 
@@ -28,24 +29,4 @@ public class LunchService {
     private void sortRecipes(List<Recipe> recipes) {
         recipesSorted = recipes; //TODO sort recipes considering best-before
     }
-
-    public List<Recipe> loadRecipes(LocalDate date) {
-        CriteriaBuilder cb = entityManager.getCriteriaBuilder();
-        CriteriaQuery<Recipe> criteriaQuery = cb.createQuery(Recipe.class);
-        Root<Recipe> recipeRoot = criteriaQuery.from(Recipe.class);
-
-        CriteriaQuery<Recipe> query = criteriaQuery.select(recipeRoot);
-
-        Subquery<Recipe> nonExpiredIngredientSubquery = query.subquery(Recipe.class);
-        Root<Recipe> nonExpiredIngredient = nonExpiredIngredientSubquery.from(Recipe.class);
-        nonExpiredIngredientSubquery.select(nonExpiredIngredient);
-
-        Predicate matchingRecipe = cb.equal(nonExpiredIngredient.get("title"), recipeRoot.get("title"));
-        Predicate expiredIngredient = cb.lessThan(nonExpiredIngredient.join("ingredients").get("useBy"), date);
-
-        Predicate allNonExpiredIngredients = cb.exists(nonExpiredIngredientSubquery.where(matchingRecipe, expiredIngredient));
-
-        return entityManager.createQuery(query.where(allNonExpiredIngredients)).getResultList();
-    }
-
 }
